Arrays
We now allow Parsing of:
foo() = 3;

We will differentiate invalid and valid assignments by parsing the argument class instead of the value if and only if the value is not needed.
If the value is need, the value is extracted and the value class is unwrapped.

An assignment is only valid if it assigns to a variable.